rename simplednn to deepset
