+++ enreg/config/models/models.yaml
@@ -2,6 +2,6 @@ defaults:
   - ParticleTransformer: ParticleTransformer
   - LorentzNet: LorentzNet
   - HPS: HPS
-  - SimpleDNN: SimpleDNN
+  - DeepSet: DeepSet
   - OmniParT: OmniParT
   - OmniDeepSet: OmniDeepSet
